+++ iOS/G3MiOSSDK/Commons/Rendererers/SGGeometryNode.cpp
@@ -37,8 +37,8 @@ void SGGeometryNode::rawRender(const G3MRenderContext* rc,
   }
 
   if (_uv != NULL) {
-    state.translateTextureCoordinates(0.0, 0.0);
-    state.scaleTextureCoordinates(1.0, 1.0);
+    state.translateTextureCoordinates((float)0.0, (float)0.0);
+    state.scaleTextureCoordinates((float)1.0, (float)1.0);
     state.setTextureCoordinates(_uv, 2, 0);
   }
